Update pre_processed_data.py
This script defines the data path as the directory containing the KITTI dataset. Then it creates a dataset object that contains the file names of all the PNG images in the data path. Next, it defines a function that loads the images and depth maps using tf.io.read_file() to read the image and depth map files, tf.image.decode_png() to decode the image, tf.image.convert_image_dtype() to convert the image dtype to float32. The depth map is loaded similarly.

It then applies the map() function to the dataset object to resize the images and depth maps using 'tf.image.resize()', and normalize the images using the formula: (image - 0.5) * 2 and depth map as well. Data augmentation is done by applying `tf.image.random_flip_left_right()'