Create Depth Estimation Monodepth Model
This script uses the Monodepth2 package to create a monocular depth estimation model. The script sets the data path to the KITTI dataset, creates a data generator, and trains the model using the fit() function. The script also defines callbacks to save the model weights and log the training progress to TensorBoard. The script saves the model and plot the training and validation loss after training.

Note that the script is a general representation and you may need to adjust the parameters according to your dataset and model. Also, you may have to adjust the path